add pop box

+++ src/app/list/list.component.ts
@@ -4,6 +4,9 @@ import { Employee } from '../employee.model';
 import { ReloadpageService} from '../reloadpage.service';
 import { Subscription } from 'rxjs/Subscription';
 import { Router,ActivatedRoute } from '@angular/router';
+import { DelboxComponent} from '../delbox/delbox.component';
+import { FilterComponent} from '../filter/filter.component';
+import {MdToolbarModule, MdDialog} from '@angular/material';
 @Component({
   selector: 'app-list',
   templateUrl: './list.component.html',
@@ -20,7 +23,9 @@ export class ListComponent implements OnInit {
     private employeeService: EmployeeService,
     private reloadpageService: ReloadpageService,
     private router: Router,
-    private activatedRoute:ActivatedRoute
+    private activatedRoute:ActivatedRoute,
+    private deldialog: MdDialog,
+    private filterdialog:MdDialog
     ) {}  
 
   ngOnInit() {
@@ -50,13 +55,22 @@ export class ListComponent implements OnInit {
       });
   }
   onDelete(){
-    this.employeeService.delete(this.selectedEmployee.id).subscribe(employee=>{  
+    let open = this.deldialog.open(DelboxComponent);
+    open.afterClosed().subscribe(result=>{
+      if(result.action=="yes"){
+          this.employeeService.delete(this.selectedEmployee.id).subscribe(employee=>{  
       this.reloadpageService.notifyOther({ option: 'delete', value: 'delete' });
       this.employeeService.get();
       this.selectedEmployee=null;
       this.router.navigate(['/add']);
       //window.location.reload();
     });
+      }
+    });
+    
+  }
+  onFilter(){
+    this.filterdialog.open(FilterComponent);
   }
   addEmployee(){
     this.selectedEmployee=null;
